Corrección CRUD producto.
