+++ Modulo_Cuentas/Modulo_Cuentas-web/src/main/webapp/Producto/producto.xhtml
@@ -38,8 +38,15 @@ and open the template in the editor.
                 <p:panel header="#{productoController.titulo} Producto" visible="#{productoController.habilitaForm}">
                     <p:panelGrid columns="2">
                         <p:outputLabel for="nombre" value="Nombre" />
-                        <p:inputTextarea id="nombre" value="#{productoController.producto.nombreProducto}"  />
+                        <p:inputTextarea id="nombre" value="#{productoController.producto.nombreProducto}" rows="2" cols="20" />
+                        <p:outputLabel for="tipoProductocmb" value="Tipo Producto" />
+                        <p:selectOneMenu id="tipoProductocmb" value="#{productoController.tipo.idTipoProducto}">
+                            <f:selectItems value="#{productoController.tiposProducto}" var="tipo" itemValue="#{tipo.idTipoProducto}" itemLabel="#{tipo.nombreTipoProducto}"/>
+                        </p:selectOneMenu>
 
+
+                        <p:outputLabel for="restriccion" value="Restriccion" />
+                        <p:inputText id="restriccion" value="#{productoController.producto.restriccionProducto}"  />
                     </p:panelGrid>
                     <p:commandButton value="Aceptar" actionListener="#{productoController.guardar}" update="@form" />
                     <p:commandButton value="Cancelar" actionListener="#{productoController.cancelar}" update="@form" />
