--- Modulo_Cuentas/Modulo_Cuentas-web/src/main/webapp/Producto/productoOri.xhtml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-To change this license header, choose License Headers in Project Properties.
-To change this template file, choose Tools | Templates
-and open the template in the editor.
--->
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:h="http://java.sun.com/jsf/html"
-      xmlns:f="http://java.sun.com/jsf/core"
-      xmlns:p="http://primefaces.org/ui"
-      xmlns:ui="http://java.sun.com/jsf/facelets">
-    <ui:composition template="./../WEB-INF/templates/plantilla.xhtml">
-
-        <ui:define name="content">
-
-            <p:panel id="mainPnl" header="Producto:" style="width: 50%; margin: auto;">
-                <p:dataTable var="producto" value="#{productoControllerOri.productos}" 
-                             selectionMode="single" selection="#{productoControllerOri.productoSel}" rowKey="#{producto.idProducto}">
-                    <p:ajax event="rowSelect" update="@form" />
-                    <p:ajax event="rowUnselect" update="@form"  />
-                    <p:column headerText="Nombre">
-                        <h:outputText value="#{producto.nombreProducto}" />
-                    </p:column>
-                    <p:column headerText="Restriccion">
-                        <h:outputText value="#{producto.restriccionProducto}" />
-                    </p:column>
-                    <p:column headerText="Tipo">
-                        <h:outputText value="#{producto.idTipoProducto.nombreTipoProducto}" />
-                    </p:column>
-                </p:dataTable>
-                <p:messages id="messages" showDetail="true" autoUpdate="true" closable="true" />
-                <p:commandButton value="Agregar" actionListener="#{productoControllerOri.agregar()}" update="@form"  />
-                <p:commandButton id="modBtn" value="Modificar" actionListener="#{productoControllerOri.modificar()}" update="@form" disabled="#{productoControllerOri.productoSel eq null}"/>
-                <p:commandButton value="Eliminar" actionListener="#{productoControllerOri.eliminar()}" update="@form" disabled="#{productoControllerOri.productoSel eq null}" >
-                     <p:confirm header="Confirmaci&#243;n" message="Desea eliminar el Producto: #{productoControllerOri.productoSel.nombreProducto}" icon="ui-icon-alert" />
-                </p:commandButton>
-                <p:panel header="#{productoControllerOri.titulo} Producto" visible="#{productoControllerOri.habilitaForm}">
-                    <p:panelGrid columns="2">
-                        <p:outputLabel for="nombre" value="Nombre" />
-                        <p:inputTextarea id="nombre" value="#{productoControllerOri.producto.nombreProducto}" rows="2" cols="20" />
-                        <p:outputLabel for="tipoProductocmb" value="Tipo Producto" />
-                        <p:selectOneMenu id="tipoProductocmb" value="#{productoControllerOri.tipo.idTipoProducto}">
-                            <f:selectItems value="#{productoControllerOri.tiposProducto}" var="tipo" itemValue="#{tipo.idTipoProducto}" itemLabel="#{tipo.nombreTipoProducto}"/>
-                        </p:selectOneMenu>
-
-
-                        <p:outputLabel for="restriccion" value="Restriccion" />
-                        <p:inputText id="restriccion" value="#{productoControllerOri.producto.restriccionProducto}"  />
-                    </p:panelGrid>
-                    <p:commandButton value="Aceptar" actionListener="#{productoControllerOri.guardar}" update="@form" />
-                    <p:commandButton value="Cancelar" actionListener="#{productoControllerOri.cancelar}" update="@form" />
-                </p:panel>
-            </p:panel>
-            <p:confirmDialog global="true" showEffect="fade" hideEffect="fade">
-                <p:commandButton value="Si" type="button" styleClass="ui-confirmdialog-yes" icon="ui-icon-check" />
-                <p:commandButton value="No" type="button" styleClass="ui-confirmdialog-no" icon="ui-icon-close" />
-            </p:confirmDialog>
-        </ui:define>
-    </ui:composition>
-</html>
