+++ Modulo_Cuentas/Modulo_Cuentas-web/src/main/webapp/WEB-INF/templates/plantilla.xhtml
@@ -26,7 +26,7 @@
                             <p:menuitem value="Ver Transacciones" url="./../Transaccion/ver.xhtml" />
                         </p:submenu>
                         <p:submenu label="Productos" icon="ui-icon-document">
-                            <p:menuitem value="Ver Productos" url="./../Producto/productoOri.xhtml" />
+                            <p:menuitem value="Ver Productos" url="./../Producto/producto.xhtml" />
 
                         </p:submenu>
                         <p:submenu label="Cuentas" icon="ui-icon-note">
